+++ config/Cargo.toml
@@ -24,6 +24,7 @@ num_enum = "0.5.1"
 rand = "0.7.3"
 rand_core = { version = "0.6.0", default-features = false }
 num_cpus = "1.10"
+parking_lot = "0.11.1"
 starcoin-types = { path = "../types" }
 starcoin-vm-types = { path = "../vm/types" }
 network-p2p-types = { path = "../network-p2p/types"}
